Moved table generation related code into TableGeneration folder. Refactored TableParameterFactory to use inheritance instead of a lot of if statements.
