+++ Dapper.Rainbow.MySql/Attributes.cs
@@ -2,30 +2,59 @@
 using System.Linq;
 using System.Reflection;
 using System.Collections.Generic;
+using Dapper.TableGeneration;
 
 namespace Dapper
 {
 	/// <summary>
-	/// The abstract Attribute class that all custom model attributes should inherit from
+	/// The abstract attribute class that all custom model attributes should inherit from
 	/// </summary>
-	public abstract class TableConstraintAttribute : Attribute { }
+	public abstract class TableAttribute : Attribute { }
+
+	/// <summary>
+	/// The abstract attribute class that all custom model attributes for Table Constraints should inherit from
+	/// </summary>
+	public abstract class TableConstraintAttribute : TableAttribute { 
+		internal abstract Constraint getConstraint(string column);
+	}
+
+	/// <summary>
+	/// The abstract attribute class that all custom model attributes for Table Modifiers should inherit from
+	/// </summary>
+	public abstract class TableModifierAttribute : TableAttribute {
+		internal abstract Modifier getModifier ();
+	}
 
 	/// <summary>
 	/// adds a primary key constraint to the property for table generation
 	/// </summary>
-	public class PrimaryKey : TableConstraintAttribute { }
+	public class PrimaryKey : TableConstraintAttribute {
+		internal override Constraint getConstraint (string column)
+		{
+			return new PrimaryKeyConstraint (column);
+		}
+ 	}
 
 	/// <summary>
 	/// adds the not null constraint to the property for table generation
 	/// </summary>
-	public class NotNull : TableConstraintAttribute { } 
+	public class NotNull : TableModifierAttribute {
+		internal override Modifier getModifier(){
+			return new NotNullModifier();
+		}
+	} 
 
 	/// <summary>
 	/// adds the auto_increment to the property in table generation
 	/// </summary>
-	public class AutoIncrement : TableConstraintAttribute { }
+	public class AutoIncrement : TableModifierAttribute { 
+		internal override Modifier getModifier ()
+		{
+			return new AutoInfrementModifier ();
+		}
+	}
 
-	internal class PrimaryKeyConstraint : Constraints
+	internal class PrimaryKeyConstraint : Constraint
 	{
 		public PrimaryKeyConstraint(string column) : base(column){ }
