+++ Test/TableParameterFactoryTests.cs
@@ -1,26 +1,25 @@
 ﻿using System;
 using PetaTest;
 using System.Collections.Generic;
+using Dapper.TableGeneration;
 
 namespace Dapper
 {
 	[TestFixture]
 	public class TableParameterFactoryTests
 	{
-		private class User {
+		private class Names {
+			[AutoIncrement]
 			[PrimaryKey]
 			public int Id { get; set; }
-		}
-
-		private class User2 {
-			[PrimaryKey]
-			public int Id { get; set; }
+			[NotNull]
 			public string Name { get; set; }
+
 		}
 
 		[Test]
 		public void ColumnsCreatedTest(){
-			var props = typeof(User2).GetProperties ();
+			var props = typeof(Names).GetProperties ();
 			var columns = new List<TableColumn> ();
 
 			foreach (var prop in props) {
@@ -32,16 +31,34 @@ namespace Dapper
 
 		[Test]
 		public void ConstraintsCreatedTest(){
-			var props = typeof(User).GetProperties();
+			var props = typeof(Names).GetProperties();
 			var columns = new List<TableColumn>();
 
 			foreach (var prop in props) {
 				columns.Add (TableParameterFactory.TableColumnFromProperty (prop));
 			}
 
-			Assert.IsTrue(columns.Count == 1);
-			Assert.IsTrue(columns [0].getConstraints().Count == 1);
+			Assert.IsTrue (hasTwoColumns (columns));
+			Assert.IsTrue (hasAConstraint (columns[0]));
+			Assert.IsTrue (hasAModifier (columns[0]));
+			Assert.IsTrue (hasAModifier (columns[1]));
 		}
+
+		static bool hasTwoColumns (List<TableColumn> columns)
+		{
+			return columns.Count == 2;
+		}
+
+		static bool hasAModifier (TableColumn column)
+		{
+			return column.getModifiers ().Count == 1;
+		}
+
+		static bool hasAConstraint (TableColumn column)
+		{
+			return column.getConstraints ().Count == 1;
+		}
+
 	}
 }
